requirements.txt
arrow==1.3.0
beautifulsoup4==4.12.3
blinker==1.7.0
bs4==0.0.2
certifi==2024.2.2
cffi==1.16.0
chardet==5.2.0
charset-normalizer==3.3.2
click==8.1.7
codecarbon==2.3.4
cryptography==42.0.5
Flask==3.0.2
Flask-Admin==1.6.1
Flask-Cors==4.0.0
Flask-Login==0.6.3
Flask-Migrate==4.0.7
Flask-SQLAlchemy==3.1.1
Flask-WTF==1.2.1
fuzzywuzzy==0.18.0
greenlet==3.0.3
idna==3.6
importlib-metadata==7.0.1
itsdangerous==2.1.2
Jinja2==3.1.3
joblib==1.3.2
jusText==3.0.0
langdetect==1.0.9
lxml==5.1.0
MarkupSafe==2.1.5
numpy==1.24.4
pandas==2.0.3
pdfminer.six==20231228
pillow==10.2.0
prometheus_client==0.20.0
psutil==5.9.8
py-cpuinfo==9.0.0
pycparser==2.21
pycryptodome==3.20.0
pynvml==11.5.0
pytest==8.3.5
python-dateutil==2.8.2
python-decouple==3.8
python-dotenv==1.0.1
pytz==2024.1
rapidfuzz==3.6.1
requests==2.31.0
scikit-learn==1.3.2
scipy==1.10.1
sentencepiece==0.2.0
six==1.16.0
soupsieve==2.5
SQLAlchemy==2.0.27
threadpoolctl==3.3.0
tqdm==4.66.2
types-python-dateutil==2.8.19.20240106
typing_extensions==4.10.0
tzdata==2024.1
urllib3==2.2.1
Werkzeug==3.0.1
WTForms==3.1.2
xmltodict==0.13.0
zipp==3.17.0
